+++ src/hocs/index.js
@@ -1,3 +1,4 @@
 export { default as withLastSearchTerms } from './withLastSearchTerms/withLastSearchTerms';
 export { default as withLocation } from './withLocation/withLocation';
 export { withUseResourcesIds } from './withUseResourcesIds/withUseResourcesIds';
+export { default as withSingleRecordImport } from './withSingleRecordImport';